+++ src/3DSecure/DataHelper.php
@@ -3,10 +3,10 @@
 /**
  * Title: Ogone 3-D Secure data helper
  * Description:
- * Copyright: Copyright (c) 2005 - 2014
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
- * @since 1.4.0
+ * @since 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Ogone_3DSecure_DataHelper {
 	/**