+++ src/Brands.php
@@ -1,0 +1,111 @@
+<?php
+
+/**
+ * Title: Ogone brands (BRAND parameter)
+ * Description:
+ * Copyright: Copyright (c) 2005 - 2015
+ * Company: Pronamic
+ * @author Remco Tolsma
+ * @version 1.1.0
+ * @since 1.1.0
+ * @see https://github.com/wp-pay-gateways/ogone/wiki/Brands
+ */
+class Pronamic_WP_Pay_Gateways_Ogone_Brands {
+	/**
+	 * Acceptgiro
+	 *
+	 * @var string
+	 */
+	const ACCEPTGIRO = 'Acceptgiro';
+
+	/**
+	 * American Express
+	 *
+	 * @var string
+	 */
+	const AMERICAN_EXPRESS = 'American Express';
+
+	/**
+	 * Bank transfer
+	 *
+	 * @var string
+	 */
+	const BANK_TRANSFER = 'Bank transfer';
+
+	/**
+	 * Bancontact/Mister Cash
+	 *
+	 * @var string
+	 */
+	const BCMC = 'BCMC';
+
+	/**
+	 * DirectEbankingNL
+	 *
+	 * @var string
+	 */
+	const DIRECT_EBANKING_NL = 'DirectEbankingNL';
+
+	/**
+	 * iDEAL
+	 *
+	 * @var string
+	 */
+	const IDEAL = 'iDEAL';
+
+	/**
+	 * Maestro
+	 *
+	 * @var string
+	 */
+	const MAESTRO = 'Maestro';
+
+	/**
+	 * MaestroUK
+	 *
+	 * @var string
+	 */
+	const MAESTRO_UK = 'MaestroUK';
+
+	/**
+	 * MasterCard
+	 *
+	 * @var string
+	 */
+	const MASTERCARD = 'MasterCard';
+
+	/**
+	 * Payment on Delivery
+	 *
+	 * @var string
+	 */
+	const PAYMENT_ON_DELIVERY = 'Payment on Delivery';
+
+	/**
+	 * PayPal
+	 *
+	 * @var string
+	 */
+	const PAYPAL = 'PAYPAL';
+
+	/**
+	 * Sofort Uberweisung
+	 *
+	 * @var string
+	 */
+	const SOFORT_UBERWEISUNG = 'Sofort Uberweisung';
+
+	/**
+	 * VISA
+	 *
+	 * @var string
+	 */
+	const VISA = 'VISA';
+
+	/**
+	 * Wallie
+	 *
+	 * @var string
+	 */
+	const WALLIE = 'Wallie';
+}