+++ src/DataCreditCardHelper.php
@@ -3,26 +3,18 @@
 /**
  * Title: Ogone data default helper class
  * Description:
- * Copyright: Copyright (c) 2005 - 2014
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
- * @since 1.4.0
+ * @version 1.1.0
+ * @since 1.0.0
  */
-class Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper {
-	/**
-	 * Data
-	 *
-	 * @var array
-	 */
-	private $data;
-
-	//////////////////////////////////////////////////
-
+class Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper extends Pronamic_WP_Pay_Gateways_Ogone_DataHelper {
 	/**
 	 * Constructs and initialize a Ogone data default helper class
 	 */
 	public function __construct( Pronamic_WP_Pay_Gateways_Ogone_Data $data ) {
-		$this->data = $data;
+		parent::__construct( $data );
 	}
 
 	//////////////////////////////////////////////////
@@ -36,9 +28,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper
 	 */
 	public function set_number( $number ) {
-		$this->data->set_field( 'CARDNO', $number );
-
-		return $this;
+		return $this->set_field( 'CARDNO', $number );
 	}
 
 	/**
@@ -48,9 +38,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper
 	 */
 	public function set_expiration_date( DateTime $date ) {
-		$this->data->set_field( 'ED', $date->format( Pronamic_WP_Pay_Gateways_Ogone_Ogone::EXPIRATION_DATE_FORMAT ) );
-
-		return $this;
+		return $this->set_field( 'ED', $date->format( Pronamic_WP_Pay_Gateways_Ogone_Ogone::EXPIRATION_DATE_FORMAT ) );
 	}
 
 	/**
@@ -60,8 +48,6 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper
 	 */
 	public function set_security_code( $security_code ) {
-		$this->data->set_field( 'CVC', $security_code );
-
-		return $this;
+		return $this->set_field( 'CVC', $security_code );
 	}
 }