+++ src/DataGeneralHelper.php
@@ -3,26 +3,18 @@
 /**
  * Title: Ogone data default helper class
  * Description:
- * Copyright: Copyright (c) 2005 - 2014
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
- * @since 1.4.0
+ * @version 1.1.0
+ * @since 1.0.0
  */
-class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
-	/**
-	 * Data
-	 *
-	 * @var Pronamic_WP_Pay_Gateways_Ogone_Data
-	 */
-	private $data;
-
-	//////////////////////////////////////////////////
-
+class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper extends Pronamic_WP_Pay_Gateways_Ogone_DataHelper {
 	/**
 	 * Constructs and initialize a Ogone data default helper class
 	 */
 	public function __construct( Pronamic_WP_Pay_Gateways_Ogone_Data $data ) {
-		$this->data = $data;
+		parent::__construct( $data );
 	}
 
 	//////////////////////////////////////////////////
@@ -36,9 +28,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_psp_id( $number ) {
-		$this->data->set_field( 'PSPID', $number );
-
-		return $this;
+		return $this->set_field( 'PSPID', $number );
 	}
 
 	//////////////////////////////////////////////////
@@ -50,9 +40,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_order_id( $order_id ) {
-		$this->data->set_field( 'ORDERID', $order_id );
-
-		return $this;
+		return $this->set_field( 'ORDERID', $order_id );
 	}
 
 	/**
@@ -63,9 +51,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_order_description( $description ) {
-		$this->data->set_field( 'COM', $description );
-
-		return $this;
+		return $this->set_field( 'COM', $description );
 	}
 
 	//////////////////////////////////////////////////
@@ -77,9 +63,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_amount( $amount ) {
-		$this->data->set_field( 'AMOUNT', Pronamic_WP_Pay_Util::amount_to_cents( $amount ) );
-
-		return $this;
+		return $this->set_field( 'AMOUNT', Pronamic_WP_Pay_Util::amount_to_cents( $amount ) );
 	}
 
 	/**
@@ -89,9 +73,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_currency( $currency ) {
-		$this->data->set_field( 'CURRENCY', $currency );
-
-		return $this;
+		return $this->set_field( 'CURRENCY', $currency );
 	}
 
 	//////////////////////////////////////////////////
@@ -103,9 +85,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_customer_name( $name ) {
-		$this->data->set_field( 'CN', $name );
-
-		return $this;
+		return $this->set_field( 'CN', $name );
 	}
 
 	/**
@@ -115,9 +95,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper
 	 */
 	public function set_email( $email ) {
-		$this->data->set_field( 'EMAIL', $email );
-
-		return $this;
+		return $this->set_field( 'EMAIL', $email );
 	}
 
 	//////////////////////////////////////////////////
@@ -129,8 +107,45 @@ class Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper {
 	 * @return Pronamic_WP_Pay_Gateways_Ogone_DirectLink_DataHelper
 	 */
 	public function set_language( $language ) {
-		$this->data->set_field( 'LANGUAGE', $language );
+		return $this->set_field( 'LANGUAGE', $language );
+	}
 
-		return $this;
+	//////////////////////////////////////////////////
+
+	/**
+	 * Set payment method
+	 *
+	 * @param string $payment_method
+	 * @return Pronamic_WP_Pay_Gateways_Ogone_DirectLink_DataHelper
+	 */
+	public function set_payment_method( $payment_method ) {
+		return $this->set_field( 'PM', $payment_method );
+	}
+
+	//////////////////////////////////////////////////
+
+	/**
+	 * Set payment methods list
+	 *
+	 * @param string $list
+	 * @return Pronamic_WP_Pay_Gateways_Ogone_DirectLink_DataHelper
+	 */
+	public function set_payment_methods_list( $list ) {
+		return $this->set_field( 'PMLIST', $list );
+	}
+
+	//////////////////////////////////////////////////
+
+	/**
+	 * Set brand of a credit/debit/purchasing card
+	 *
+	 * If you send the BRAND field without sending a value in the PM field (‘CreditCard’ or ‘Purchasing Card’),
+	 * the BRAND value will not be taken into account.
+	 *
+	 * @param string $brand
+	 * @return Pronamic_WP_Pay_Gateways_Ogone_DirectLink_DataHelper
+	 */
+	public function set_brand( $brand ) {
+		return $this->set_field( 'BRAND', $brand );
 	}
 }