+++ src/DirectLink/Config.php
@@ -3,10 +3,10 @@
 /**
  * Title: Ogone DirectLink config
  * Description:
- * Copyright: Copyright (c) 2005 - 2011
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
- * @version 1.0
+ * @version 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Ogone_DirectLink_Config extends Pronamic_WP_Pay_Gateways_Ogone_Config {
 	public $user_id;