+++ src/OrderStandard/Client.php
@@ -3,10 +3,10 @@
 /**
  * Title: Ogone order standard client
  * Description:
- * Copyright: Copyright (c) 2005 - 2014
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
- * @version 1.0
+ * @version 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Client {
 	/**