+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## 1.1.0
+*	Added constant class for the Ogone BRAND parameter.
+*	Added constant class for the Ogone PM (payment method) parameter/
+*	Simplified the Ogone data helper classes.
+*	Added an payment methods list class for the Ogone PMLIST parameter.
+*	Added direct support for the Bancontact/Mister Cash payment method.
+
 ## 1.0.0
 *	First release.